# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Tether-ERP Procurement API
  description: |
    Comprehensive REST API for procurement management system with approval workflows,
    budget constraints, document linking, and multi-stage approval chains.
  version: 1.0.0
  contact:
    name: Tether-ERP Support
    email: support@tether-erp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development Server
  - url: https://api.tether-erp.com/api/v1
    description: Production Server
...
paths:
...
  /categories/{id}/budget-codes:
    get:
      tags:
        - Categories
      summary: Get category budget codes
      operationId: getCategoryBudgetCodes
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of budget codes
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags:
        - Categories
      summary: Add budget code to category
      operationId: addCategoryBudgetCode
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - budgetCode
              properties:
                budgetCode:
                  type: string
      responses:
        "201":
          description: Budget code added
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Resource not found